dashboard/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
singlestoredb>=1.4.0
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
